requirements.txt
python-docx==1.0.0
PyQt5==5.15.9
python-dotenv==1.0.0
pandas==2.0.3